+++ GameServer/ECS-Effects/StagECSEffect.cs
@@ -49,7 +49,7 @@ namespace DOL.GS
 				m_amount = (int)(OwnerPlayer.CalculateMaxHealth(OwnerPlayer.Level, OwnerPlayer.GetModified(eProperty.Constitution)) * m_amountPercent);
 			else m_amount = (int)(Owner.MaxHealth * m_amountPercent);
 
-			Owner.BaseBuffBonusCategory[(int)eProperty.MaxHealth] += m_amount;
+			Owner.BaseBuffBonusCategory[eProperty.MaxHealth] += m_amount;
 			Owner.Health += (int)(Owner.GetModified(eProperty.MaxHealth) * m_amountPercent);
 			if (Owner.Health > Owner.MaxHealth) Owner.Health = Owner.MaxHealth;
 
@@ -65,7 +65,7 @@ namespace DOL.GS
         {
 			Owner.Model = m_originalModel;
 			
-			Owner.BaseBuffBonusCategory[(int)eProperty.MaxHealth] -= m_amount;
+			Owner.BaseBuffBonusCategory[eProperty.MaxHealth] -= m_amount;
 			if (Owner.IsAlive && Owner.Health > Owner.MaxHealth)
 				Owner.Health = Owner.MaxHealth;
