+++ GameServer/propertycalc/EvadeChanceCalculator.cs
@@ -40,14 +40,14 @@ namespace DOL.GS.PropertyCalc
                 if (player.HasAbility(Abilities.Evade))
                     chance += (int) ((((player.Dexterity + player.Quickness) / 2 - 50) * 0.05 + player.GetAbilityLevel(Abilities.Evade) * 5) * 10);
 
-                chance += player.BaseBuffBonusCategory[(int) property] * 10;
-                chance += player.SpecBuffBonusCategory[(int) property] * 10;
-                chance -= player.DebuffCategory[(int) property] * 10;
-                chance += player.OtherBonus[(int) property] * 10;
-                chance += player.AbilityBonus[(int) property] * 10;
+                chance += player.BaseBuffBonusCategory[property] * 10;
+                chance += player.SpecBuffBonusCategory[property] * 10;
+                chance -= player.DebuffCategory[property] * 10;
+                chance += player.OtherBonus[property] * 10;
+                chance += player.AbilityBonus[property] * 10;
             }
             else if (living is GameNPC npc)
-                chance += npc.AbilityBonus[(int)property] * 10 + npc.EvadeChance * 10;
+                chance += npc.AbilityBonus[property] * 10 + npc.EvadeChance * 10;
 
             return chance;
         }
