+++ GameServer/realmabilities/effects/BarrierOfFortitudeEffect.cs
@@ -42,7 +42,7 @@ namespace DOL.GS.Effects
 
 			GameEventMgr.AddHandler(m_player, GamePlayerEvent.Quit, new DOLEventHandler(PlayerLeftWorld));
 
-			m_player.AbilityBonus[(int)eProperty.ArmorAbsorption] += m_value;
+			m_player.AbilityBonus[eProperty.ArmorAbsorption] += m_value;
 
 			m_player.EffectList.Add(this);
 			player.TempProperties.SetProperty(RealmAbilities.BarrierOfFortitudeAbility.BofBaSb, true);
@@ -73,7 +73,7 @@ namespace DOL.GS.Effects
 		{
 
 			StopTimers();
-			m_player.AbilityBonus[(int)eProperty.ArmorAbsorption] -= m_value;
+			m_player.AbilityBonus[eProperty.ArmorAbsorption] -= m_value;
 			m_player.EffectList.Remove(this);
 			GameEventMgr.RemoveHandler(m_player, GamePlayerEvent.Quit, new DOLEventHandler(PlayerLeftWorld));
 			m_player.TempProperties.RemoveProperty(RealmAbilities.BarrierOfFortitudeAbility.BofBaSb);
